// sdks/typescript/src/generated/Client.ts
export declare namespace WiseClient {
    export type Options = BaseClientOptions;

    export interface RequestOptions extends BaseRequestOptions {}
}

export class WiseClient {
// ...
    ): Promise<Response> {
        return core.makePassthroughRequest(
            input,
            init,
            {
                baseUrl:
                    this._options.baseUrl ??
                    (async () => {
                        const env = await core.Supplier.get(this._options.environment);
                        return typeof env === "string" ? env : (env as Record<string, string>)?.api;
                    }),
                headers: this._options.headers,
                timeoutInSeconds: this._options.timeoutInSeconds,
                maxRetries: this._options.maxRetries,
                fetch: this._options.fetch,
                logging: this._options.logging,
                getAuthHeaders: async () => (await this._options.authProvider.getAuthRequest()).headers,
            },
            requestOptions,
        );
    }
}