// sdks/typescript/src/generated/BaseClient.ts
export type AuthOption =
    | false
    | core.AuthProvider["getAuthRequest"]
    | core.AuthProvider
    | AnyAuthProvider.AuthOptions<[BearerAuthProvider.AuthOptions, BasicAuthProvider.AuthOptions]>;

export type BaseClientOptions = {
    environment: core.Supplier<environments.WiseEnvironment | environments.WiseEnvironmentUrls>;
    /** Specify a custom URL to connect the client to. */
    baseUrl?: core.Supplier<string>;
    /** Override the X-External-Correlation-Id header */
    externalCorrelationId?: core.Supplier<string | undefined>;
    /** Additional headers to include in requests. */
    headers?: Record<string, string | core.Supplier<string | null | undefined> | null | undefined>;
    /** The default maximum time to wait for a response in seconds. */
    timeoutInSeconds?: number;
    /** The default number of times to retry the request. Defaults to 2. */
    maxRetries?: number;
    /** Provide a custom fetch implementation. Useful for platforms that don't have a built-in fetch or need a custom implementation. */
    fetch?: typeof fetch;
    fetcher?: core.FetchFunction;
    /** Configure logging for the client. */
    logging?: core.logging.LogConfig | core.logging.Logger;
    /** Default options for SSE stream reconnection behavior. Has no effect on non-resumable endpoints. */
    stream?: { reconnectionEnabled?: boolean; maxReconnectionAttempts?: number };
    /** Override auth. Pass false to disable, a function returning auth headers, an AuthProvider, or auth options. */
    auth?: AuthOption;
} & AnyAuthProvider.AuthOptions<[BearerAuthProvider.AuthOptions, BasicAuthProvider.AuthOptions]>;