// sdks/typescript/src/generated/environments.ts
// This file was auto-generated by Fern from our API Definition.

export interface WiseEnvironmentUrls {
    api: string;
    cards: string;
}

export const WiseEnvironment = {
    Sandbox: {
        api: "https://api.wise-sandbox.com/2026Q3",
        cards: "https://twcard.wise-sandbox.com",
    },
    Production: {
        api: "https://api.wise.com/2026Q3",
        cards: "https://twcard.wise.com",
    },
    SandboxMtls: {
        api: "https://api-mtls.wise-sandbox.com/2026Q3",
        cards: "https://twcard.wise-sandbox.com",
    },
    ProductionMtls: {
        api: "https://api-mtls.transferwise.com/2026Q3",
        cards: "https://twcard.wise.com",
    },
} as const;

export type WiseEnvironment =
    | typeof WiseEnvironment.Sandbox
    | typeof WiseEnvironment.Production
    | typeof WiseEnvironment.SandboxMtls
    | typeof WiseEnvironment.ProductionMtls;